Refactor filter/reject/where into FilterOperations module; further refactoring is going to get messy
