+++ lib/cascading.rb
@@ -15,6 +15,7 @@ require 'cascading/flow'
 require 'cascading/mode'
 require 'cascading/operations'
 require 'cascading/identity_operations'
+require 'cascading/filter_operations'
 require 'cascading/regex_operations'
 require 'cascading/text_operations'
 require 'cascading/scope'
